+++ src/dto/site.js
@@ -28,6 +28,7 @@ export const SiteDto = {
     const siteData = {
       id: jsonObject.id,
       baseURL: jsonObject.baseURL,
+      hlxConfig: jsonObject.hlxConfig,
       deliveryType: jsonObject.deliveryType,
       gitHubURL: jsonObject.gitHubURL,
       organizationId: jsonObject.organizationId,
@@ -57,6 +58,7 @@ export const SiteDto = {
   toJSON: (site) => ({
     id: site.getId(),
     baseURL: site.getBaseURL(),
+    hlxConfig: site.getHlxConfig(),
     deliveryType: site.getDeliveryType(),
     gitHubURL: site.getGitHubURL(),
     organizationId: site.getOrganizationId(),