+++ test/controllers/hooks.test.js
@@ -25,7 +25,7 @@ const { expect } = chai;
 
 function getExpectedSlackMessage(baseURL, channel, source, hlxConfig) {
   const cdnConfigPart = hlxConfig
-    ? `, _HLX Version_: *5*, _Dev URL_: https://${hlxConfig.rso.ref}--${hlxConfig.rso.site}--${hlxConfig.rso.owner}.aem.live`
+    ? `, _HLX Version_: *5*, _Dev URL_: \`https://${hlxConfig.rso.ref}--${hlxConfig.rso.site}--${hlxConfig.rso.owner}.aem.live\``
     : '';
   return Message()
     .channel(channel)
@@ -285,8 +285,6 @@ describe('Hooks Controller', () => {
       context.dataAccess.upsertSiteCandidate.resolves();
 
       const expectedConfig = {
-        cdnProdHost: undefined,
-        domain: 'some-domain.com',
         hlxVersion: 4,
         rso: {},
       };
@@ -299,16 +297,104 @@ describe('Hooks Controller', () => {
 
       const expectedMessage = {
         channel: 'channel-id',
+        unfurl_links: false,
         blocks: [
           {
             text: {
               type: 'mrkdwn',
-              text: 'HLX config updated for existing site: *<https://some-domain.com|https://some-domain.com>*, _HLX Version_: *4*, _Dev URL_: https://undefined--undefined--undefined.aem.live',
+              text: 'HLX config added for existing site: *<https://some-domain.com|https://some-domain.com>*, _HLX Version_: *4*, _Dev URL_: `https://undefined--undefined--undefined.aem.live`',
             },
             type: 'section',
           },
         ],
       };
+
+      const resp = await (await hooksController.processCDNHook(context)).json();
+      expect(resp).to.equal('CDN site candidate disregarded');
+      expect(context.dataAccess.updateSite.calledOnce).to.be.true;
+      expect(
+        context.dataAccess.updateSite.firstCall.args[0].getHlxConfig(),
+      ).to.deep.equal(expectedConfig);
+      expect(context.slackClients.WORKSPACE_INTERNAL_STANDARD.postMessage.calledOnce).to.be.true;
+      expect(
+        context.slackClients.WORKSPACE_INTERNAL_STANDARD.postMessage.firstCall.args[0],
+      ).to.deep.equal(expectedMessage);
+      expect(context.log.warn).to.have.been.calledWith('Could not process site candidate. Reason: Site candidate already exists in sites db, Source: CDN, Candidate: https://some-domain.com');
+    });
+
+    it('while candidate is disregarded, hlx config is updated if different from site', async () => {
+      context.dataAccess.siteCandidateExists.resolves(false);
+      context.dataAccess.upsertSiteCandidate.resolves();
+      context.data = {
+        forwardedHost: 'some-domain.com, main--some-site--some-owner.hlx.live',
+      };
+
+      const hlxConfig = {
+        cdn: { prod: { host: 'some-domain.com' } },
+        code: {},
+        content: {
+          title: 'helix-website',
+          contentBusId: 'another-id',
+          source: {
+            type: 'google',
+            url: 'https://drive.google.com/drive/u/3/folders/abcd1234',
+            id: '5678',
+          },
+        },
+        hlxVersion: 5,
+      };
+
+      const expectedConfig = {
+        ...hlxConfig,
+        rso: {
+          ref: 'main',
+          site: 'some-site',
+          owner: 'some-owner',
+          tld: 'hlx.live',
+        },
+      };
+
+      context.dataAccess.getSiteByBaseURL.resolves(SiteDto.fromJson({
+        baseURL: 'https://some-domain.com',
+        isLive: true,
+        deliveryType: 'aem_edge',
+        hlxConfig: {
+          cdn: { prod: { host: 'some-cdn-host.com' } },
+          content: {
+            title: 'helix-website',
+            contentBusId: 'fooid',
+            source: {
+              type: 'google',
+              url: 'https://drive.google.com/drive/u/3/folders/16251625162516',
+              id: '1234',
+            },
+          },
+          hlxVersion: 5,
+          rso: {
+            ref: 'main',
+            site: 'some-site',
+            owner: 'some-owner',
+          },
+        },
+      }));
+
+      const expectedMessage = {
+        channel: 'channel-id',
+        unfurl_links: false,
+        blocks: [
+          {
+            text: {
+              type: 'mrkdwn',
+              text: 'HLX config updated for existing site: *<https://some-domain.com|https://some-domain.com>*, _HLX Version_: *5*, _Dev URL_: `https://main--some-site--some-owner.aem.live`',
+            },
+            type: 'section',
+          },
+        ],
+      };
+
+      nock('https://admin.hlx.page')
+        .get('/config/some-owner/aggregated/some-site.json')
+        .reply(200, hlxConfig);
 
       const resp = await (await hooksController.processCDNHook(context)).json();
       expect(resp).to.equal('CDN site candidate disregarded');